Switch activity dao count method to return a long, also updated dependent methods. Part of #1363.

+++ core/src/main/java/edu/northwestern/bioinformatics/studycalendar/dao/ActivityDao.java
@@ -171,8 +171,8 @@ public class ActivityDao extends StudyCalendarMutableDomainObjectDao<Activity> i
     }
 
     @SuppressWarnings({ "unchecked" })
-    public Integer getCount() {
-        return  (Integer) CollectionUtils.firstElement(
+    public Long getCount() {
+        return (Long) CollectionUtils.firstElement(
                 getSession().createCriteria(Activity.class)
                     .setProjection( Projections.projectionList()
                     .add(Projections.rowCount()))
